Properly close the applications.list file
buildFavorites wasn't closing explicitly the favorites file fd
after reading the whole file. This is not a critical issue, since
the garbage collector would eventually close it by itself, but
the file is reopened for writing in the subsequent
self.saveFavorites() call while the object holding the open file
descriptors is still in scope and thus not eligible for GC.

Also, the path generation code for the applications.list file has
been refactored into the __init__ method of the plugin, so that it
remains consistent across the functions that use it.