Add timeout to search entry focus
This fixes an issue where the pointer steals focus of another
application within the same application list whenever the user is
typing. The timeout ensures that the first element is selected
regardless of the position of the pointer.

At this point, it seems more like a hack than anything, but it seems to
fix the behavior. I'm starting to think a refactoring is due...