Remove search focus delay
When opening the menu, the search text input gets focused after a 100ms
delay. This was done to avoid the mouse stealing focus when the pointer
was in the way of the menu prior to opening. However, it meant that if
the user starts typing immediately after opening the menu, your search
input focus would be lost.

This is now solved by only allowing the pointer to focus on buttons
after it has actually moved over them after entering. Now the user
should be able to start typing their application name immediately,
making the whole search menu experience more responsive.